+++ src/main.ts
@@ -1,9 +1,14 @@
 import { StableDiffusion } from "./stable-diffusion";
-import { mkdir } from "fs/promises";
 import { join } from "path";
 import { GeneratedImage } from "./models";
 import { TaskQueue } from "./task-queue";
-import { CleanupTask, IndexTask, ResizeTask, SdTask } from "./tasks";
+import {
+  CleanupTask,
+  CreateDirectoryTask,
+  IndexTask,
+  ResizeTask,
+  SdTask,
+} from "./tasks";
 import { Presets, SingleBar } from "cli-progress";
 import { AxiosError } from "axios";
 import { Configuration } from "./configuration";
@@ -31,12 +36,7 @@ import { Configuration } from "./configuration";
     )
   );
 
-  const progress = new SingleBar(
-    {
-      etaAsynchronousUpdate: true,
-    },
-    Presets.shades_classic
-  );
+  const progress = new SingleBar({}, Presets.shades_classic);
   progress.start(taskQueue.length, 0);
   taskQueue.on(TaskQueue.TASK_ERROR, (task, err) => {
     console.error(`Task failed: ${err}`);
@@ -75,8 +75,8 @@ async function buildTaskQueue(
 
   for (const [category, styles] of Object.entries(modifiers)) {
     for (const style of styles) {
-      const folder = join(__dirname, "..", "outputs", category, style);
-      await mkdir(folder, { recursive: true });
+      const folder = join(cfg.outputPath, category, style);
+      taskQueue.enqueue(new CreateDirectoryTask(folder));
 
       for (const input of inputs) {
         const name = cfg.nameToPrompt(input.name);
@@ -85,7 +85,7 @@ async function buildTaskQueue(
           instance,
           prompt,
           {
-            initialImagePath: join(__dirname, "..", "inputs", input.name),
+            initialImagePath: join(cfg.inputPath, input.name),
             ...options,
           },
           (index) => join(folder, `${name}-${index}-full.png`)