+++ src/tasks/index.ts
@@ -1,5 +1,6 @@
 export * from "./task";
 export * from "./cleanup-task";
+export * from "./create-directory-task";
 export * from "./index-task";
 export * from "./resize-task";
 export * from "./sd-task";